Add file manifest to Releaser configuration (#179)
* releaser: modify build.sh to output a 'file manifest' for comparison between builds

This manifest can be used to compare a build to a previous build, in order to validate any assumptions about what headers and library files are installed by 'make install'. 

* circleci: stop using sourceforge mirror for downloading Windows pcre dependency

The existing mirror disappeared. We can use sourceforge's download URL instead of a direct mirror,
as long as the -UserAgent "NativeHost" flag is passed to the Invoke-WebRequest commandlet (otherwise it downloads a web page.)