Compile 3rd party object libs as position-independent code when building shared library
It appears that CMake doesn't add the -fPIC flag to object libraries that are added to shared libraries (i.e. ldserverapi when BUILD_SHARED_LIBS=ON). See https://gitlab.kitware.com/cmake/cmake/-/issues/17356. 

This should theoretically cause build failures - and it does - on certain platforms, namely the Lua SDKs.

I'm not clear why the cmake tests pass, and Releaser is able to generate shared libraries for the project without issue.

It's possible that older version of CMake + GCC cause an interaction where the flag needs to be manually added. More investigation is needed to get to the bottom of this.